pyproject.toml
[project]
name = "conversational-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.5.1",
    "fastapi>=0.135.2",
    "google-genai>=0.4.0",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
